requirements.txt
packaging==23.2
uvicorn==0.27.1
fastapi== 0.109.2
python-dotenv==1.0.1
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
langchain-openai==0.0.6
pydantic==1.10.14